perf: remove builder container
Building the nix shell is time consuming and having to build it twice,
consecutively is unreasonable. This may slightly increase the image size
but who really cares.

This also brings the production environment inline with the development
and ci environments. Which is the final piece needed to ensure a
completely unified and consistent environment for the entire SDLC
(awesome!).